ASYLUM: HiveControl puzzle: fix a bug
_leverPosition - _prevLeverPosition should be signed.

+++ engines/asylum/puzzles/hivecontrol.cpp
@@ -87,7 +87,7 @@ void PuzzleHiveControl::reset() {
 
 	_frameIndexes[kElementSwirlRim] = 0;
 	if (_leverPosition != _prevLeverPosition) {
-		_leverDelta = (uint32)abs((double)(_leverPosition - _prevLeverPosition)) * 16 / 5;
+		_leverDelta = (uint32)abs((double)_leverPosition - (double)_prevLeverPosition) * 16 / 5;
 		_currentControl = kControlGlyph4;
 	}
 }
@@ -283,7 +283,7 @@ bool PuzzleHiveControl::mouseLeftDown(const AsylumEvent &) {
 	case kControlGlyph5:
 	case kControlGlyph6:
 		_leverPosition = _currentControl - 49;
-		_leverDelta = (uint32)abs((double)(_leverPosition - _prevLeverPosition)) * (GraphicResource::getFrameCount(_vm, getWorld()->graphicResourceIds[kElementLever]) - 1) / 5;
+		_leverDelta = (uint32)abs((double)_leverPosition - (double)_prevLeverPosition) * (GraphicResource::getFrameCount(_vm, getWorld()->graphicResourceIds[kElementLever]) - 1) / 5;
 		if (_leverDelta)
 			getSound()->playSound(getWorld()->graphicResourceIds[76], false, Config.sfxVolume - 10);
 	}
